+++ src/main/java/lucie/hitchhike/Hitchhike.java
@@ -1,13 +1,11 @@
 package lucie.hitchhike;
 
 import lucie.hitchhike.item.ItemAlias;
-import lucie.hitchhike.item.ItemContent;
+import lucie.hitchhike.item.ItemHorse;
 import lucie.hitchhike.item.ItemPouch;
-import lucie.hitchhike.item.content.ItemHorse;
-import lucie.hitchhike.item.content.ItemSkeletonHorse;
-import lucie.hitchhike.item.content.ItemZombieHorse;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -16,10 +14,6 @@ import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 @Mod(Hitchhike.MODID)
 public class Hitchhike
@@ -30,9 +24,6 @@ public class Hitchhike
     // The mod logger.
     public static final Logger LOGGER = LogManager.getFormatterLogger("Hitchhike");
 
-    // List of all pouches with mobs.
-    public static final List<ItemContent> HORSES = Arrays.asList(ItemAlias.POUCH_WITH_HORSE, ItemAlias.POUCH_WITH_SKELETON_HORSE, ItemAlias.POUCH_WITH_ZOMBIE_HORSE);
-
     public Hitchhike()
     {
         // Used for texture overrides.
@@ -41,8 +32,23 @@ public class Hitchhike
 
     private void clientSetup(final FMLClientSetupEvent event)
     {
-        // Add texture override for horse models.
-        event.enqueueWork(() -> ItemProperties.register(ItemAlias.POUCH_WITH_HORSE, new ResourceLocation("hitchhike", "model"), (stack, level, living, id) -> ItemHorse.getModel(stack)));
+        // Add override for 'pouch_with_horse' model.
+        event.enqueueWork(() -> ItemProperties.register(ItemAlias.POUCH_WITH_HORSE, new ResourceLocation("hitchhike", "model"), (stack, level, living, identification) ->
+        {
+            // Get data from compound.
+            if (stack.getTag() != null && stack.getTag().contains("Content"))
+            {
+                // Get id from variant.
+                int id = (stack.getTag().getCompound("Content").getInt("Variant")%8) + 1;
+
+                // Convert id to model id.
+                return 0.0000001F * id;
+            }
+
+            // Return error model.
+            return 0.0000000F;
+        }));
+
     }
 
     @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
@@ -55,9 +61,9 @@ public class Hitchhike
             event.getRegistry().register(new ItemPouch());
 
             // Pouch Content.
-            event.getRegistry().register(new ItemHorse());
-            event.getRegistry().register(new ItemSkeletonHorse());
-            event.getRegistry().register(new ItemZombieHorse());
+            event.getRegistry().register(new ItemHorse(EntityType.HORSE));
+            event.getRegistry().register(new ItemHorse(EntityType.SKELETON_HORSE));
+            event.getRegistry().register(new ItemHorse(EntityType.ZOMBIE_HORSE));
         }
     }
 }